# openapi/specs/system.json
{"components":{"schemas":{"AnnouncementReadDTO":{"properties":{"announcementId":{"format":"int64","type":"integer"}},"type":"object"},"AnnouncementVO":{"properties":{"announcement":{"$ref":"#/components/schemas/SysAnnouncement"},"hasUnReadAnnouncement":{"format":"int32","type":"integer"}},"type":"object"},"ApiStatisticsCreateDTO":{"properties":{"endTime":{"format":"date-time","type":"string"},"startTime":{"format":"date-time","type":"string"},"statisticsInterval":{"format":"int32","type":"integer"},"type":{"format":"int32","type":"integer"}},"required":["endTime","startTime","statisticsInterval","type"],"type":"object"},"CreateResEntity":{"properties":{"id":{"format":"int64","type":"integer"}},"type":"object"},"CreateUserDTO":{"properties":{"email":{"type":"string"},"nickname":{"type":"string"},"password":{"type":"string"},"phoneNumber":{"type":"string"},"sex":{"format":"int32","type":"integer"},"username":{"type":"string"}},"required":["nickname","password","username"],"type":"object"},"GetSysPermissionRuleDTO":{"properties":{"permissionRuleName":{"type":"string"}},"type":"object"},"ImportFailUser":{"properties":{"reason":{"type":"string"},"username":{"type":"string"}},"type":"object"},"IncreaseApiUsageDTO":{"properties":{"time":{"format":"date-time","type":"string"},"type":{"format":"int32","type":"integer"}},"required":["time","type"],"type":"object"},"KnowledgeBaseImportUser":{"properties":{"className":{"type":"string"},"createBy":{"format":"int64","type":"integer"},"createTime":{"format":"date-time","type":"string"},"nickname":{"type":"string"},"params":{"additionalProperties":{"type":"object"},"type":"object"},"password":{"type":"string"},"remark":{"type":"string"},"updateBy":{"format":"int64","type":"integer"},"updateTime":{"format":"date-time","type":"string"},"userId":{"format":"int64","type":"integer"},"username":{"type":"string"}},"type":"object"},"KnowledgeBaseUserImportDTO":{"properties":{"failUserList":{"items":{"$ref":"#/components/schemas/ImportFailUser"},"type":"array"},"knowledgeBaseImportUserList":{"items":{"$ref":"#/components/schemas/KnowledgeBaseImportUser"},"type":"array"}},"required":["knowledgeBaseImportUserList"],"type":"object"},"KnowledgeBaseUserVO":{"properties":{"nickname":{"type":"string"},"permissions":{"format":"int32","type":"integer"},"userId":{"format":"int64","type":"integer"},"username":{"type":"string"}},"type":"object"},"LoginUser":{"properties":{"ipaddr":{"type":"string"},"loginTime":{"format":"int64","type":"integer"},"longTerm":{"type":"boolean"},"permissions":{"items":{"type":"string"},"type":"array","uniqueItems":true},"role":{"type":"string"},"sysUser":{"$ref":"#/components/schemas/SysUser"},"token":{"$ref":"#/components/schemas/Token"},"userId":{"format":"int64","type":"integer"},"username":{"type":"string"}},"type":"object"},"PageBeanKnowledgeBaseUserVO":{"properties":{"current":{"format":"int32","type":"integer"},"pages":{"format":"int32","type":"integer"},"rows":{"items":{"$ref":"#/components/schemas/KnowledgeBaseUserVO"},"type":"array"},"total":{"format":"int64","type":"integer"}},"type":"object"},"PageBeanSysUser":{"properties":{"current":{"format":"int32","type":"integer"},"pages":{"format":"int32","type":"integer"},"rows":{"items":{"$ref":"#/components/schemas/SysUser"},"type":"array"},"total":{"format":"int64","type":"integer"}},"type":"object"},"PublicUserInfoVO":{"properties":{"avatar":{"type":"string"},"nickname":{"type":"string"},"userId":{"format":"int64","type":"integer"},"username":{"type":"string"}},"type":"object"},"ResDataAnnouncementVO":{"properties":{"code":{"type":"string"},"data":{"$ref":"#/components/schemas/AnnouncementVO"},"msg":{"type":"string"}},"type":"object"},"ResDataCreateResEntity":{"properties":{"code":{"type":"string"},"data":{"$ref":"#/components/schemas/CreateResEntity"},"msg":{"type":"string"}},"type":"object"},"ResDataInteger":{"properties":{"code":{"type":"string"},"data":{"format":"int32","type":"integer"},"msg":{"type":"string"}},"type":"object"},"ResDataKnowledgeBaseUserImportDTO":{"properties":{"code":{"type":"string"},"data":{"$ref":"#/components/schemas/KnowledgeBaseUserImportDTO"},"msg":{"type":"string"}},"type":"object"},"ResDataListSysApiStatisticsVO":{"properties":{"code":{"type":"string"},"data":{"items":{"$ref":"#/components/schemas/SysApiStatisticsVO"},"type":"array"},"msg":{"type":"string"}},"type":"object"},"ResDataListSysConfig":{"properties":{"code":{"type":"string"},"data":{"items":{"$ref":"#/components/schemas/SysConfig"},"type":"array"},"msg":{"type":"string"}},"type":"object"},"ResDataLoginUser":{"properties":{"code":{"type":"string"},"data":{"$ref":"#/components/schemas/LoginUser"},"msg":{"type":"string"}},"type":"object"},"ResDataLong":{"properties":{"code":{"type":"string"},"data":{"format":"int64","type":"integer"},"msg":{"type":"string"}},"type":"object"},"ResDataPageBeanKnowledgeBaseUserVO":{"properties":{"code":{"type":"string"},"data":{"$ref":"#/components/schemas/PageBeanKnowledgeBaseUserVO"},"msg":{"type":"string"}},"type":"object"},"ResDataPageBeanSysUser":{"properties":{"code":{"type":"string"},"data":{"$ref":"#/components/schemas/PageBeanSysUser"},"msg":{"type":"string"}},"type":"object"},"ResDataPublicUserInfoVO":{"properties":{"code":{"type":"string"},"data":{"$ref":"#/components/schemas/PublicUserInfoVO"},"msg":{"type":"string"}},"type":"object"},"ResDataString":{"properties":{"code":{"type":"string"},"data":{"type":"string"},"msg":{"type":"string"}},"type":"object"},"ResDataSysOrganization":{"properties":{"code":{"type":"string"},"data":{"$ref":"#/components/schemas/SysOrganization"},"msg":{"type":"string"}},"type":"object"},"ResDataSysPermissionRule":{"properties":{"code":{"type":"string"},"data":{"$ref":"#/components/schemas/SysPermissionRule"},"msg":{"type":"string"}},"type":"object"},"ResDataSysUser":{"properties":{"code":{"type":"string"},"data":{"$ref":"#/components/schemas/SysUser"},"msg":{"type":"string"}},"type":"object"},"ResetPasswordDTO":{"properties":{"newPassword":{"maxLength":15,"minLength":6,"type":"string"},"userId":{"format":"int64","type":"integer"}},"required":["newPassword","userId"],"type":"object"},"SysAnnouncement":{"properties":{"attachment":{"type":"string"},"author":{"type":"string"},"content":{"type":"string"},"createBy":{"format":"int64","type":"integer"},"createTime":{"format":"date-time","type":"string"},"datePublished":{"format":"date-time","type":"string"},"deleted":{"format":"int32","type":"integer"},"id":{"format":"int64","type":"integer"},"params":{"additionalProperties":{"type":"object"},"type":"object"},"pinned":{"format":"int32","type":"integer"},"remark":{"type":"string"},"title":{"type":"string"},"type":{"type":"string"},"updateBy":{"format":"int64","type":"integer"},"updateTime":{"format":"date-time","type":"string"}},"type":"object"},"SysApiStatisticsListDTO":{"properties":{"endTime":{"format":"date-time","type":"string"},"startTime":{"format":"date-time","type":"string"},"statisticsInterval":{"format":"int32","type":"integer"},"type":{"format":"int32","type":"integer"}},"required":["endTime","startTime","statisticsInterval","type"],"type":"object"},"SysApiStatisticsVO":{"properties":{"createTime":{"format":"date-time","type":"string"},"endTime":{"format":"date-time","type":"string"},"id":{"format":"int64","type":"integer"},"startTime":{"format":"date-time","type":"string"},"statisticsInterval":{"format":"int32","type":"integer"},"type":{"format":"int32","type":"integer"},"updateTime":{"format":"date-time","type":"string"},"usageCount":{"format":"int32","type":"integer"}},"type":"object"},"SysConfig":{"properties":{"description":{"type":"string"},"id":{"format":"int64","type":"integer"},"name":{"type":"string"},"value":{"type":"string"}},"type":"object"},"SysOrganization":{"properties":{"ancestors":{"type":"string"},"createBy":{"format":"int64","type":"integer"},"createTime":{"format":"date-time","type":"string"},"deleted":{"format":"int32","type":"integer"},"email":{"type":"string"},"id":{"format":"int64","type":"integer"},"leader":{"type":"string"},"orderNum":{"format":"int32","type":"integer"},"organizationName":{"type":"string"},"params":{"additionalProperties":{"type":"object"},"type":"object"},"parentId":{"format":"int64","type":"integer"},"phone":{"type":"string"},"remark":{"type":"string"},"status":{"format":"int32","type":"integer"},"updateBy":{"format":"int64","type":"integer"},"updateTime":{"format":"date-time","type":"string"}},"type":"object"},"SysPermissionRule":{"properties":{"associationTableName":{"type":"string"},"createBy":{"format":"int64","type":"integer"},"createTime":{"format":"date-time","type":"string"},"deleteTime":{"format":"int64","type":"integer"},"deleted":{"format":"int32","type":"integer"},"entityIdFieldName":{"type":"string"},"entityTableName":{"type":"string"},"id":{"format":"int64","type":"integer"},"knowledgeBaseAssociationType":{"format":"int32","type":"integer"},"knowledgeBaseIdFieldName":{"type":"string"},"params":{"additionalProperties":{"type":"object"},"type":"object"},"permissionRuleName":{"type":"string"},"permissionsFieldName":{"type":"string"},"remark":{"type":"string"},"requirePermission":{"format":"int32","type":"integer"},"updateBy":{"format":"int64","type":"integer"},"updateTime":{"format":"date-time","type":"string"},"userAssociated":{"format":"int32","type":"integer"},"userAssociatedTableName":{"type":"string"}},"type":"object"},"SysRole":{"properties":{"createBy":{"format":"int64","type":"integer"},"createTime":{"format":"date-time","type":"string"},"dataScope":{"format":"int32","type":"integer"},"deleted":{"format":"int32","type":"integer"},"id":{"format":"int64","type":"integer"},"params":{"additionalProperties":{"type":"object"},"type":"object"},"remark":{"type":"string"},"roleKey":{"type":"string"},"roleName":{"type":"string"},"roleSort":{"format":"int64","type":"integer"},"status":{"type":"string"},"updateBy":{"format":"int64","type":"integer"},"updateTime":{"format":"date-time","type":"string"}},"type":"object"},"SysUser":{"properties":{"avatar":{"type":"string"},"createBy":{"format":"int64","type":"integer"},"createTime":{"format":"date-time","type":"string"},"deleted":{"format":"int32","type":"integer"},"email":{"type":"string"},"id":{"format":"int64","type":"integer"},"loginDate":{"format":"date-time","type":"string"},"loginIp":{"type":"string"},"nickname":{"type":"string"},"organizations":{"items":{"$ref":"#/components/schemas/SysOrganization"},"type":"array"},"params":{"additionalProperties":{"type":"object"},"type":"object"},"password":{"type":"string"},"phoneNumber":{"type":"string"},"remark":{"type":"string"},"role":{"$ref":"#/components/schemas/SysRole"},"sex":{"format":"int32","type":"integer"},"status":{"format":"int32","type":"integer"},"updateBy":{"format":"int64","type":"integer"},"updateTime":{"format":"date-time","type":"string"},"username":{"type":"string"}},"type":"object"},"Token":{"properties":{"accessToken":{"type":"string"},"accessTokenExpirationTime":{"format":"int64","type":"integer"},"refreshToken":{"type":"string"}},"type":"object"}},"securitySchemes":{"Authorization":{"in":"header","name":"Authorization","type":"apiKey"}}},"info":{"contact":{"email":"yxlmzch0222@outlook.com","name":"Anynote","url":"https://github.com/zch0222/anynote"},"description":"Anynote 平台后端 API 文档","license":{"name":"Apache 2.0","url":"https://www.apache.org/licenses/LICENSE-2.0"},"termsOfService":"https://github.com/zch0222/anynote","title":"Anynote API","version":"0.5.1"},"openapi":"3.0.1","paths":{"/":{"get":{"operationId":"getSysConfigs_1","responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResDataListSysConfig"}}},"description":"OK"}},"tags":["系统配置"]}},"/announcements/latest":{"get":{"operationId":"getLatestAnnouncement","responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResDataAnnouncementVO"}}},"description":"OK"}},"tags":["系统公告"]}},"/announcements/read":{"post":{"operationId":"readAnnouncement","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnnouncementReadDTO"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResDataCreateResEntity"}}},"description":"OK"}},"tags":["系统公告"]}},"/apiStatistics":{"post":{"operationId":"createApiStatistics","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiStatisticsCreateDTO"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResDataLong"}}},"description":"OK"}},"tags":["API统计"]}},"/apiStatistics/getApiStatistics":{"post":{"operationId":"getApiStatistics","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SysApiStatisticsListDTO"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResDataListSysApiStatisticsVO"}}},"description":"OK"}},"tags":["API统计"]}},"/apiStatistics/increaseUsage":{"post":{"operationId":"increaseUsage","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IncreaseApiUsageDTO"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResDataString"}}},"description":"OK"}},"tags":["API统计"]}},"/hello":{"get":{"operationId":"helloWorld","responses":{"200":{"content":{"*/*":{"schema":{"type":"string"}}},"description":"OK"}},"tags":["test-controller"]}},"/organizations/bases/info/{organizationId}":{"get":{"operationId":"getOrganizationInfoByKnowledgeBaseId","parameters":[{"in":"path","name":"organizationId","required":true,"schema":{"format":"int64","type":"integer"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResDataSysOrganization"}}},"description":"OK"}},"tags":["组织管理"]}},"/permissionRules":{"get":{"operationId":"getSysPermissionRule","parameters":[{"in":"query","name":"getSysPermissionRuleDTO","required":true,"schema":{"$ref":"#/components/schemas/GetSysPermissionRuleDTO"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResDataSysPermissionRule"}}},"description":"OK"}},"tags":["权限规则"]}},"/permissionRules/{id}":{"get":{"operationId":"getSysPermissionRuleById","parameters":[{"in":"path","name":"id","required":true,"schema":{"format":"int64","type":"integer"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResDataSysPermissionRule"}}},"description":"OK"}},"tags":["权限规则"]}},"/user":{"post":{"operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDTO"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResDataCreateResEntity"}}},"description":"OK"}},"tags":["用户管理"]}},"/user/bases":{"get":{"operationId":"getKnowledgeBaseUsers","parameters":[{"in":"query","name":"knowledgeBaseId","required":true,"schema":{"format":"int64","type":"integer"}},{"in":"query","name":"page","required":true,"schema":{"format":"int32","type":"integer"}},{"in":"query","name":"pageSize","required":true,"schema":{"format":"int32","type":"integer"}},{"in":"query","name":"username","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResDataPageBeanKnowledgeBaseUserVO"}}},"description":"OK"}},"tags":["用户管理"]}},"/user/bases/import":{"post":{"operationId":"importKnowledgeBaseUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/KnowledgeBaseUserImportDTO"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResDataKnowledgeBaseUserImportDTO"}}},"description":"OK"}},"tags":["用户管理"]}},"/user/info/{username}":{"get":{"operationId":"getUserInfo","parameters":[{"in":"path","name":"username","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResDataLoginUser"}}},"description":"OK"}},"tags":["用户管理"]}},"/user/list":{"get":{"operationId":"getManageUserList","parameters":[{"in":"query","name":"page","required":true,"schema":{"format":"int32","type":"integer"}},{"in":"query","name":"pageSize","required":true,"schema":{"format":"int32","type":"integer"}},{"in":"query","name":"username","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResDataPageBeanSysUser"}}},"description":"OK"}},"tags":["用户管理"]}},"/user/mine":{"get":{"description":"前端登录态自检与初次拉取用户资料","operationId":"getMyInfo","responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResDataSysUser"}}},"description":"OK"}},"summary":"获取当前登录用户信息","tags":["用户管理"]}},"/user/pubInfo/{username}":{"get":{"operationId":"getPublicUserInfo_1","parameters":[{"in":"path","name":"username","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResDataSysUser"}}},"description":"OK"}},"tags":["用户管理"]}},"/user/publicUserInfo":{"get":{"operationId":"getPublicUserInfo","parameters":[{"in":"query","name":"userId","required":true,"schema":{"format":"int64","type":"integer"}},{"in":"query","name":"username","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResDataPublicUserInfoVO"}}},"description":"OK"}},"tags":["用户管理"]}},"/user/resetPassword":{"post":{"operationId":"resetPassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDTO"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResDataString"}}},"description":"OK"}},"tags":["用户管理"]}},"/user/{userId}":{"get":{"operationId":"getSysUserById","parameters":[{"in":"path","name":"userId","required":true,"schema":{"format":"int64","type":"integer"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResDataSysUser"}}},"description":"OK"}},"tags":["用户管理"]},"put":{"operationId":"updateSysUser","parameters":[{"in":"path","name":"userId","required":true,"schema":{"format":"int64","type":"integer"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SysUser"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResDataInteger"}}},"description":"OK"}},"tags":["用户管理"]}},"/user/{userId}/ban":{"post":{"operationId":"banUser","parameters":[{"in":"path","name":"userId","required":true,"schema":{"format":"int64","type":"integer"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResDataString"}}},"description":"OK"}},"tags":["用户管理"]}},"/user/{userId}/info":{"get":{"operationId":"getSysUserInfoById","parameters":[{"in":"path","name":"userId","required":true,"schema":{"format":"int64","type":"integer"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResDataSysUser"}}},"description":"OK"}},"tags":["用户管理"]}},"/user/{userId}/unban":{"post":{"operationId":"unBanUser","parameters":[{"in":"path","name":"userId","required":true,"schema":{"format":"int64","type":"integer"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResDataString"}}},"description":"OK"}},"tags":["用户管理"]}}},"security":[{"Authorization":[]}],"tags":[{"description":"系统配置管理接口","name":"系统配置"},{"description":"系统公告管理","name":"系统公告"},{"description":"权限规则管理接口","name":"权限规则"},{"description":"组织架构管理接口","name":"组织管理"},{"description":"系统API调用统计","name":"API统计"},{"description":"系统用户管理接口","name":"用户管理"}]}