# services/auth/src/main/java/com/anynote/auth/controller/TokenController.java
package com.anynote.auth.controller;

import com.anynote.auth.model.dto.LoginDTO;
import com.anynote.auth.model.dto.LoginRequestDTO;
import com.anynote.auth.model.dto.LogoutDTO;
import com.anynote.auth.model.dto.RefreshTokenDTO;
import com.anynote.auth.model.dto.RegisterDTO;
import com.anynote.auth.model.dto.ResetPasswordDTO;
import com.anynote.auth.service.LoginService;
import com.anynote.auth.service.impl.LoginServiceImpl;
import com.anynote.common.security.utils.SecurityUtils;
import com.anynote.core.utils.ResUtil;
import com.anynote.core.web.model.bo.ResData;
import com.anynote.system.api.model.bo.LoginUser;
import com.anynote.system.api.model.bo.Token;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RestController;

import jakarta.validation.Valid;

/**
 * 认证 Controller
 * @author 称霸幼儿园
 */
@Tag(name = "认证", description = "登录、重置密码、注册")
@RestController
@Validated
public class TokenController {

    @Autowired
    private LoginService loginService;

    @Operation(summary = "用户登录")
    @PostMapping("login")
    public ResData<LoginDTO> login(@RequestBody LoginRequestDTO loginRequestDTO) {
        LoginUser loginUser = loginService.login(loginRequestDTO.getUsername(), loginRequestDTO.getPassword());
        return ResData.success(new LoginDTO(loginUser));
    }

    @Operation(summary = "重置密码")
    @PostMapping("resetPassword")
    public ResData<LoginDTO> resetPassword(@RequestBody @Valid ResetPasswordDTO resetPasswordDTO) {
        LoginUser loginUser = loginService.resetPassword(resetPasswordDTO);
        return ResData.success(new LoginDTO(loginUser));
    }

    @Operation(summary = "健康探测", description = "返回 OK；运维与前端探活使用")
    @GetMapping("test")
    public ResData<String> test() {
        return ResData.success("OK");
    }

    @Operation(summary = "用户注册")
    @PostMapping("register")
    public ResData<LoginDTO> register(@RequestBody @Valid RegisterDTO registerDTO) {
        return ResUtil.success(new LoginDTO(loginService.register(registerDTO)));
    }

    @Operation(summary = "刷新 Token",
            description = "使用 refreshToken 换取新的 access + refresh 对，旧 refreshToken 立即失效。供 BFF 在 accessToken 过期时调用。")
    @PostMapping("refresh")
    public ResData<Token> refresh(@RequestBody @Valid RefreshTokenDTO refreshTokenDTO) {
        return ResData.success(loginService.refresh(refreshTokenDTO.getRefreshToken()));
    }

    @Operation(summary = "登出",
            description = "清除当前会话 accessToken / refreshToken 的 Redis 缓存。幂等：token 已过期或不存在时静默成功。")
    @PostMapping("logout")
    public ResData<Void> logout(@RequestBody @Valid LogoutDTO logoutDTO) {
        loginService.logout(logoutDTO.getAccessToken(), logoutDTO.getRefreshToken());
        return ResData.success(null);
    }

//    public static void main(String[] args) {
//        System.out.println(SecurityUtils.encryptPassword("123456"));
//    }
}